+++ src/main/java/com/sprint/mission/discodeit/dto/auth/LoginRequestDto.java
@@ -1,0 +1,6 @@
+package com.sprint.mission.discodeit.dto.auth;
+
+public record LoginRequestDto(
+        String email,
+        String password
+) {}
