+++ src/main/java/com/sprint/mission/discodeit/dto/auth/LoginResponseDto.java
@@ -1,0 +1,11 @@
+package com.sprint.mission.discodeit.dto.auth;
+
+import java.util.UUID;
+
+public record LoginResponseDto(
+        UUID id,
+        String email,
+        String name,
+        boolean isOnline,
+        UUID profileImage
+) {}
