+++ src/main/java/com/sprint/mission/discodeit/dto/binarycontent/FileUploadDto.java
@@ -1,0 +1,8 @@
+package com.sprint.mission.discodeit.dto.binarycontent;
+
+public record FileUploadDto(
+        String fileName,
+        String contentType,
+        byte[] content,
+        long fileSize
+) {}
