+++ src/main/java/com/sprint/mission/discodeit/dto/channel/ChannelCreateDto.java
@@ -1,14 +1,11 @@
-package com.sprint.mission.discodeit.dto;
+package com.sprint.mission.discodeit.dto.channel;
 
 import com.sprint.mission.discodeit.entity.ChannelType;
 import org.springframework.lang.Nullable;
 
-public class ChannelDto {
-
-    public record Create(
+public record ChannelCreateDto(
         String name,
         ChannelType type,
         @Nullable String topic,
         @Nullable String description
-    ) {}
-}
+) {}
