+++ src/main/java/com/sprint/mission/discodeit/dto/readstatus/ChannelUnreadStatusDto.java
@@ -1,0 +1,8 @@
+package com.sprint.mission.discodeit.dto.readstatus;
+
+import java.util.UUID;
+
+public record ChannelUnreadStatusDto(
+        UUID channelId,
+        boolean hasUnread
+) {}
